Started Work On "Room Value" Context Thing, 10/2/2022
Started Work On "Room Value" Context Thing. That will attempt to estimate the value of a room that a lootpile is in, then use this to modify or determine if a chest should replace said lootpile, and if so what type of chest it should be, what rarity, and value, etc, 10/2/2022.