Got Some Work Done On Rolling Dice To Determine If A Chest Should Spawn Or Not, 10/11/2022
Got Some Work Done On Rolling Dice To Determine If A Chest Should Spawn Or Not. Also did some planning for tomorrow on how I might do determining what materials the chest and lock should be if a chest is generated, using a "hat full of tickets" method of some form, 10/11/2022.