Added These "ItemGroupOdds" To Each Dungeon Type That Will Be Used To Determine Loot Chances Later During Generation, 11/12/2022
Added These "ItemGroupOdds" To Each Dungeon Type That Will Be Used To Determine Loot Chances Later During Generation. This is really not an implementation that I have much confidence in, but I had to get something down, otherwise I was literally going to be spinning my wheels forever trying to think of a "perfect" way to do this, so for now this is just what it is until I find some better way possibly later on in testing, etc, 11/12/2022.