Continued On Current Method For Loot Generation, This Time Material Modifier Values To Item Odds, 12/23/2022
Continued On Current Method For Loot Generation, This Time Material Modifier Values To Item Odds. Finally decided to get back on the modding wagon for this, somewhat atleast. Still not 100% sold on this current method of loot generation, but at this point I'll have to see it fail to really consider overhauling it entirely first, so just continuing with it for now. With some placeholder chest material modifiers in place for the loot odds, next I'll probably try to add in the room modifier values as well, then maybe with these modified item odds actually start rolling some items and see how it goes, for now though, just going slow to get back into it. Oh yeah, don't forget about all the other stuff of course like material of the lock and other stuff to modify the loot odds and such, if that makes the cut in the end, right now it's still too much feature creep that I need to not let it slow me down so much before just getting a working prototype atleast going first, 12/23/2022.