Got A Bit More of The Loot Generation Stuff In Place, This Time For Mostly Some Armor Item Types, 12/30/2022
Got A Bit More of The Loot Generation Stuff In Place, This Time For Mostly Some Armor Item Types. Plan to do the weapon types next I work on this, will be a bit more weird due to some values I have in those particular enums, such as wands, which should not have a material and stuff like that, will see when I get there, 12/30/2022.