Started Work On A "miscGroupOdds" Array For Other Left Over Loot Chances, 1/4/2023
Started Work On A "miscGroupOdds" Array For Other Left Over Loot Chances. Now this is probably not the best way to do this, but I think how much worse can it get, ha. But will continue working on those values for the dungeon types, then actually add the "work" part of adding these items and such, also applying condition damage modifiers to appropriate items in the end. Oh yeah, also made a temporary "solution" for the stacking supplies that are unstackable in vanilla like the bandages, will just deal with the specific mod related stuff involving that later, or let it deal with itself hopefully based on the check used, etc, 1/4/2023.