Finished Basics For "miscGroupOdds" Array, 1/7/2023
Finished Basics For "miscGroupOdds" Array, and how items should spawn due to the values from it, for now atleast, needs alot of work from what it currently is, but good placeholder for now basically. Also added in Condition Modifier for various items based on the dungeon type spawned in. Next I work on this I'll likely want to do testing for this new added loot feature addition, after that I'm still not sure what should be worked on next though, will have to see, 1/7/2023.