Bit More Happy With Lock-Pick Formula, Still Needs Testing And Tweaking, 1/12/2023
Bit More Happy With Lock-Pick Formula, Still Needs Testing And Tweaking. It's not perfect but good enough for now. I plan on working on the actual logic next for success/failure of picking the lock, such as counting up the attributes on the chest, stuff like that. I'll probably work on the "lock jamming" related stuff later as that will have its own formula fun most likely, so will see, 1/12/2023.