Got Good Amount of Lock-Picking Logic And Formula Done, Also Started Work On Bashing Formula And Logic, 1/15/2023
Got Good Amount of Lock-Picking Logic And Formula Done, Also Started Work On Bashing Formula And Logic. Got a surprising amount of stuff done today, was not expecting to get much at all done initially. But yeah, now Lock-Picking is mostly good for the time being, same with Lock Jamming chances and odds, also got some work done on similar formulas and logic for the bashing behaviors, so good progress to continue on with next time, 1/15/2023.