More Work On Lock Bashing Formula, Still Needs Logic For Stuff Later, 1/18/2023
More Work On Lock Bashing Formula, Still Needs Logic For Stuff Later. Will likely skip the logic part for the actually bashing and consequences to weapon durability and such, to instead next work on the actual rolls and formula for if the lock should break after being bashed, then maybe the stuff for hitting the chest body itself after that, will see, 1/18/2023.