Made Good Progress On The Chest Bashing Formula Today, 1/21/2023
Made Good Progress On The Chest Bashing Formula Today. Got most of the physical bashing formula stuff in a place I'm happy enough with for now. Next time I work on this I'll likely start work on the logic part involving this specifically, like damaging the player health, fatigue, weapon durability from bashing and such. Also if I get past that part and find a good place to put it all, I'll likely work on the actual condition damage part for the chest contents based on all the bashing stuff and how that will all work and such, 1/21/2023.