Some Framework And Start Of Formulas For "BashingCost" Logic/Methods, 1/22/2023
Some Framework And Start Of Formulas For "BashingCost" Logic/Methods. Will want to optimize and cleanup the math and variables done so far in this method currently being used to damage player health/fatigue/weapon condition from the action of physically bashing a chest. But will worry about the cleanup part after I have the stuff mostly in place for the many permutations that will be in this method. I'm positive I can reduce down many of the formulas that I have done already for this, but for the time being it just made more sense as I was trying to figure things out, give it time basically, 1/22/2023.