Completed Remaining Bashing Logic And Formulas Today, 2/4/2023
Completed Remaining Bashing Logic And Formulas Today. Those being the rest of the stuff for destruction spells thrown at a chest, as well as arrows from the player's bow. While still need testing, they appear to be in a state to be functional and tested atleast, hopefully. With getting those done, decided to next work on some of the junk loot replacement stuff, mostly to give certain items more "junk/scraps" depending on the item in question being destroyed/replaced. So like a piece of chest armor will on average give alot more scraps of whatever material than say a dagger of the same material, etc. Or books giving more paper scraps than a single sheet of a map that sort of thing. I'll finish assigning these values next I work on this, then after that I'll have to think of what I should next work on, getting closer and closer to being ready to go public I'd say. Still a bunch of stuff to be done though, such as adding custom sprites, audio-files, and all that sort of thing, 2/4/2023.