Mostly Version Control Commit, 2/6/2023
Mostly Version Control Commit. Got the values entered in for the scrap amounts for the weapons and armor and such, so I'd say I'm done with that part for the time being. I STILL need to do testing and such, but I really don't feel like doing it, so I'm just going to continue on and work on the chest "Inspection" formulas, rolls, logic, and possibly text based on that and such, and maybe after that finally do some actual testing and bug fixing and such, will see, 2/6/2023.