Got Much of The Roll Related Formula Done For Chest Inspection, 2/7/2023
Got Much of The Roll Related Formula Done For Chest Inspection. Still more work to do on that, but seems like it's in a pretty decent spot right now. So next I have to work on the text and how to display said text and such. After some consideration, I think I'm going to try and do something similar to what I did for the "Viewable Skill Progress" mod, where I kind of used a custom text-box/panel to make my own sort of text formatting space that was fairly easy to work with after being set-up. So I'll see about copying that from VSP next I work on this and see how it goes, 2/7/2023.