Just Did Some Testing With Adding Custom Sound Files Mostly, 2/23/2023
Just Did Some Testing With Adding Custom Sound Files Mostly. Will refine this more next time, but for now was mostly just seeing if it would as I expected at all. Also need to remember to test if the "spam click" exploit was fixed for the open-effect, forgot to check this session, not a big deal wither way, 2/23/2023.