Fixed A Few Audio Related Bugs And Sound Level, 2/28/2023
Fixed A Few Audio Related Bugs And Sound Level. Some sounds are not as loud as I'd like them to be, and for some reason changing the volume modifier in code does not seem to be changing some of them, not a big deal but will try to eventually fix that at some point. Right now this commit is just mostly a few fixes after some testing with the new audio clips. Commit right after this will be just a big vomit of files, nothing interesting. Next I work on this though, I am considering working on the save-data related stuff finally, but will see if I'm up to that or not, will see, 2/28/2023.