More Fiddling With Save Data, Getting Closer Though, 3/5/2023
More Fiddling With Save Data, Getting Closer Though. So it seems right now the loot and most of the important stuff about the chests is being saved and loading in afterward as well. Issue right now is that chests not within the "main-block" of a dungeon are being given the wrong parent object basically, and thus are loaded in the void instead of in their proper dungeon block they originally were a child to. Still trying to figure out how I'm going to deal with this specifically, also probably some other bugs as I keep testing, but good progress still. Also the save-files are fairly large even with just a few chests saved at a time, so not sure how I'll deal with that if it's even possible. Also need to figure out how to keep the replaced loot-piles from respawning after loading the game, maybe something that when they are emptied by the player they are not destroyed but set inactive instead or something? Not sure, 3/5/2023.