Made A Bunch of Changes To The Save-Data Stuff, Needs Testing Still, 3/8/2023
Made A Bunch of Changes To The Save-Data Stuff, Needs Testing Still. I'm somewhat confident that I have most of the save/load stuff solved for both the closed and "open" chests, that being the custom objects and the loot-pile objects. Still needs testing either way, but also especially for the attempted new solution to trying to deal with the "regenerating" static random loot-piles, but will see next time how it all goes, hopefully this will be all that's needed for now to move onto stuff again like balancing/fixing the formulas and other bugs, etc, 3/8/2023.