More Work Done With Reworking Bashing To HP Based, 3/12/2023
More Work Done With Reworking Bashing To HP Based. Next try and determine the actual Hit Point values of the lock and chest body, as well as for the "lock mechanism" that being the HP for the lock-jamming basically. With this, also determine how bashing might damage the lock mech. HP along-side the other bashing HP damage. Also want to not forget the magic damage related refactoring for "Damage HP" spell effects atleast thrown at the chest and such, 3/12/2023.