Outside Testing Think I'm Done With The Hit Point Based Refactoring, 3/12/2023
Outside Testing Think I'm Done With The Hit Point Based Refactoring. Still need to test if the values and how it all works feels "right" in-game and such, but in terms of code logic and ball-park values, I think I'm happy to move onto the next thing on the to-do list. That likely going to be trying to fix the loot-damage stuff from bashing and such to make it more "punishing" to some extent and more frequently destroy items and other things, 3/12/2023.