Changed A Bunch Of Values From Loot Damage System, 3/12/2023
Changed A Bunch Of Values From Loot Damage System. Still need to test all of the changes to see if I'm happy with the result and such, but I'm certain it's better than it was before atleast. Next I work on this I'll probably try to get the loot generation stuff out of just "placeholder" status and try to get it in a spot I'm more looking for, will see what happens, slow but steady progress as usual, 3/12/2023.