Added Sound Test Console Command And Some Fixes, 3/20/2023
Added Sound Test Console Command And Some Fixes. The new sound test console command is actually very useful, so that was worth doing. Found that some of the previous "bugs" I observed were not actually related to the sounds necessarily, like the occasional not playing of a bash sound when bashing was actually the bash just not registering something when spam attacking with certain attack animations with weapons, not certain why but it's not a big deal so I'll just leave that as is. Changed the volume of a couple sound-clips, but decided not to go much further when I realized exporting them caused them to get like 20-40% larger for some reason, no clue why, but just decided for now it's good enough so whatever. Now that I sort of swept the audio "concerns" under the rug for now, I'm going to probably start working on getting the chests to spawn and work inside building interiors and such, houses, etc. Also might first try and get the 3D model chests to be replaced 100% of the time with a LLC chest, will see. 3/20/2023.