Finally Got Chests Generating In Non-Dungeon Interiors, 3/22/2023
Finally Got Chests Generating In Non-Dungeon Interiors. Still need to polish them and their general logic stuff, but for now it is functional which is great. Next work on the crime/rep reduction related logic, fix the bug when a chest replaces a 3D model that it floats in the air, potentially have chests spawn in-place of others objects as well such as crates with a reduced chance or something, work on the loot-tables for the individual building types, and bug testing and fixing obviously, so far so good though, 3/22/2023.