Think I'm Done With Crime Detection And House Stuff For Now, 3/24/2023
Think I'm Done With Crime Detection And House Stuff For Now. Got most of the basic crime detection and registration stuff done for the time being, did some testing and it appears to work as I'd generally expect (did not test the thieves or DB's versions though.) Also fixed some bugs/oversights involving the chests spawning in non-dungeon building interiors. Still needs more testing as always and some issues like the Y-position, but I seemed to mostly have the Y-position thing in a "good enough." state for the time being. With that all said, very nice progress today, I think what I'll work on next is probably trying to fill out the non-dungeon interiors "loot-tables" and get those in a better place, then after that part is in a satisfactory spot maybe attempt to use that 3D model from the Unity Asset store that I was looking at? I won't spend too much time with that if it's not going smoothly, I'll just skip that and move to the next thing, which is honestly getting pretty close to ready for first release I'd say, good stuff, 3/24/2023.