Added Loot Tables For Non-Dungeons, Also Some Loot Changes, 3/26/2023
Added Loot Tables For Non-Dungeons, Also Some Loot Changes. Have the loot-tables completed for the different building types and guildhall types. Also made some rewords to the material generation system for loot again, still not perfect but definitely better. And finally am getting how to use 3D models and Prefabs, so tomorrow I might try to get the 3D model "option" to be a thing and fix whatever quirks comes with that. Also try to fix the 2D versions of the sprites to loot better and fix floating pixel on the base chest sprite and stuff like that, also other bug fixing stuff, etc, 3/26/2023.