Bunch Of Work On Implementing 3D Chest Models, 3/26/2023
Bunch Of Work On Implementing 3D Chest Models. Surprisingly good progress on this aspect, but still need to do a bunch of stuff like probably implementing 3D piles for the "destroyed" chests and such, also tomorrow probably first work on getting the save-data stuff working for the 3D models and such, also will work on a bunch of settings for these various things involving the chest graphics, etc, 3/26/2023.