Got Some Work Done On Implementing Model And Sprite Graphics, 3/27/2023
Got Some Work Done On Implementing Model And Sprite Graphics. Still need to replace a bunch of the things from other scripts and methods so they know if they should create a model or a sprite. Also with that, I'll have to consider how I'm going to deal with all this in a similar way for the "Smashed" and "Disintegrated" chest graphic variants, will just have to see when I get there I suppose, also testing to make sure saving and loading works for all this, etc, 3/27/2023.