Commit After Hard-drive Started To Seemingly Die, 3/29/2023
Commit After Hard-drive Started To Seemingly Die. So the night before this commit was made, my hard-drive started to show symptoms (again) of dying. So decided to make an emergency backup and commit this on a another PC that I could actually launch Github Desktop. So hopefully none of these files were corrupted, but will have to see, just keep note of it, 3/29/2023.