Added Decent Amount of Text Variants For Chest Interactions, 4/4/2023
Added Decent Amount of Text Variants For Chest Interactions. I think this is definitely sufficient for now atleast. Also just a note, this is my first commit being done on my "new machine" that is currently using Linux as the OS instead of Windows, so will see if there are any issues as we continue related to that, hopefully not, so far so good. Still need to test if these text variants work in-game, but I'd imagine they will. Will probably try to work on the loot and chest generation stuff again next, will see, 4/4/2023.