Even More Progress On Loot Odds Method Refactoring, 4/15/2023
Even More Progress On Loot Odds Method Refactoring. Hopefully tomorrow I'll be able to just say "good enough" for this particular system and move onto the final touches and formula tweaking stuff. But main thing I need to do tomorrow for this refactoring thing is keep tinkering with the material selection and loot gen. in general, but major part is going to be changing the "base item odds" for all the locations and dungeon types and such, to nerf some and buff others, will take a bit of time to change all/most of those values, but will just see tomorrow and go from there, 4/15/2023.