Finally Said "Good Enough" For The Loot System, 4/16/2023
Finally Said "Good Enough" For The Loot System. I finally hit the breaks on messing with the loot system, it is definitely not perfect or even up to my overly high arbitrary standards for it, but it is "Good Enough" for the initial release atleast. With that, I'm going to now work on the tweaking of the formulas and such, will do this tomorrow and then hopefully move onto the final steps and boring stuff for releasing the mod publically, finally, 4/16/2023.