Slight Tweaks To Magic Bashing Formula To Make Easier, 5/24/2025
Slight Tweaks To Magic Bashing Formula To Make Easier. After a few complaints and my own testing, tried to change some of the numbers involving the chest bashing using magic spells to be a bit more forgiving, so hopefully it can be a more effective/viable option to use, will see with more testing and feedback, 5/24/2025.